feature/achievements - Roll out homepage
