--- Layouts/index.js
@@ -1,2 +1,0 @@
-export { default as ComingSoon } from './ComingSoon'
-export { default as Homepage } from './Homepage'
